Add methods to ProveedorController for retrieving a single Proveedor and its Productos; update Proveedor model to include Productos list

+++ src/main/java/com/farmacia/repository/LoteRepository.java
@@ -11,7 +11,7 @@ import com.farmacia.models.Lote;
 public interface LoteRepository extends JpaRepository<Lote, Long> {
 
 
-    public List<Lote> findByProductoId(Integer idProducto);
+    public List<Lote> findByProductoId(Long idProducto);
 
     public Lote findByProductoIdAndLote(Integer idProducto, String lote);
